+++ grass-diary/src/state/user/useUser.ts
@@ -11,28 +11,6 @@ import { semantic } from '@styles/semantic';
 import { INTERACTION } from '@styles/interaction';
 import { useModal } from '@state/modal/useModal';
 import { useNetwork } from '@hooks/useNetwork';
-import { jwtDecode } from 'jwt-decode';
-import { useAuthActions } from '@state/auth/authStore';
-
-const getTokenExpirationDate = (token: string) => {
-  try {
-    const decode = jwtDecode(token);
-    if (!decode.exp) return;
-    const date = new Date(0);
-    date.setUTCSeconds(decode.exp);
-    return date;
-  } catch (error) {
-    console.error('Failed to decode token:', error);
-    return null;
-  }
-};
-
-const hasExpired = (expirationDate: Date) => {
-  if (!expirationDate) return false;
-  const currentDate = new Date();
-  const timeDiff = expirationDate.getTime() - currentDate.getTime();
-  return timeDiff <= 0 ? true : false;
-};
 
 const fetchAxios = async () => {
   const res = await API.get(END_POINT.member_info);
@@ -40,9 +18,7 @@ const fetchAxios = async () => {
 };
 
 export const useUser = () => {
-  const accessToken = localStorage.getItem('accessToken');
   const { isAuthenticated } = useAuth();
-  const { setIsAuthenticated } = useAuthActions();
   const memberId = useMemberId();
   const setMemberId = useSetMemberId();
   const { modal } = useModal();
@@ -92,44 +68,6 @@ export const useUser = () => {
       }
     }
   }, [isAuthenticated, isError, isSuccess]);
-
-  // 로그인 만료 타이머
-  useEffect(() => {
-    let expireInterval: NodeJS.Timeout;
-
-    if (accessToken) {
-      const expirationDate = getTokenExpirationDate(accessToken);
-
-      if (expirationDate) {
-        expireInterval = setInterval(() => {
-          if (hasExpired(expirationDate)) {
-            const setting = {
-              title: MODAL.authentication_error.title,
-              content: MODAL.authentication_error.content,
-            };
-
-            const button1 = {
-              active: true,
-              text: MODAL.confirm,
-              color: semantic.light.accent.solid.hero,
-              interaction: INTERACTION.accent.subtle(),
-              clickHandler: () => (window.location.href = '/'),
-            };
-
-            localStorage.removeItem('accessToken');
-            localStorage.setItem('logout', 'true');
-            setIsAuthenticated(false);
-            setMemberId(0);
-            clearInterval(expireInterval);
-            modal(setting, button1);
-          }
-        }, 60000);
-      }
-    }
-    return () => {
-      if (expireInterval) clearInterval(expireInterval);
-    };
-  }, [accessToken]);
 
   return memberId;
 };
